Fixed failure to handle spaces or other special characters when running exes via mono.
In the file path for ProcessHelper.platformInAction, psi.FileName
needs to be surrounded in quotes so that if the filename of the exe
has a space in it, the call to "mono <filename>" does not fail.

Simple surrounded psi.FileName with quotes, which should handle
all special characters and whitespace.